Added new hitlet submodule. Updated peak_splitting to allow hitlet splititng.
