+++ strax/__init__.py
@@ -30,3 +30,4 @@ from strax.processing.peak_building import *
 from strax.processing.peak_merging import *
 from strax.processing.peak_splitting import *
 from strax.processing.peak_properties import *
+from strax.processing.hitlets import *
